feat: comprehensive revamp — 6 bug fixes, Superset dashboards, Hex deepening, Deepnote archived
Bug Fixes:
- Deribit interest_8h normalization: /100 → *100 (rates were 100x understated)
- mart_venue_comparison: wide → long format (venue_long/venue_short/spread_bps)
- Cumulative yield: daily-compounded formula (was 365x overstated)
- APY: spread_bps/100.0 (was double-annualized *3.65)
- Hourly queries: timestamp → hour_start column
- arb_apy_pct: computed from data (was hardcoded 0)

Infrastructure:
- Apache Superset Docker Compose (localhost:8088)
- Supabase PostgreSQL connection via IPv6-enabled bridge
- 4 dashboards, 14 charts created via REST API

Analysis:
- Hex report: effect sizes (Cohen's d), normality checks (Shapiro-Wilk)
- Cointegration test (Engle-Granger), Sharpe ratio, transaction cost modeling

Documentation:
- README with Superset setup, Data Corrections section
- Deepnote notebooks archived to archive/deepnote/

+++ sql/annualized-yield/05_yield_by_symbol.sql
@@ -12,7 +12,7 @@ SELECT
     symbol,
     asset_class,
     ROUND(AVG(daily_annualized_yield_pct), 4) AS avg_annualized_yield_pct,
-    ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS median_yield_pct,
+    ROUND((PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS median_yield_pct,
     ROUND(MAX(daily_annualized_yield_pct), 4) AS max_yield_pct,
     ROUND(MIN(daily_annualized_yield_pct), 4) AS min_yield_pct,
     COUNT(DISTINCT date) AS total_days
