+++ sql/annualized-yield/08_yield_quartiles.sql
@@ -10,15 +10,11 @@
 
 SELECT
     venue,
-    ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS q1_yield_pct,
-    ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS q2_yield_pct,
-    ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS q3_yield_pct,
-    ROUND(PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS q4_yield_pct,
-    ROUND(
-        PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY daily_annualized_yield_pct)
-        - PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY daily_annualized_yield_pct),
-        4
-    ) AS iqr_pct
+    ROUND((PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS q1_yield_pct,
+    ROUND((PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS q2_yield_pct,
+    ROUND((PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS q3_yield_pct,
+    ROUND((PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS q4_yield_pct,
+    ROUND((PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY daily_annualized_yield_pct) - PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS iqr_pct
 FROM marts.mart_daily_funding
 GROUP BY venue
 ORDER BY q2_yield_pct DESC;
