+++ sql/annualized-yield/12_yield_seasonality.sql
@@ -13,7 +13,7 @@ SELECT
     EXTRACT(MONTH FROM date)::int AS month_num,
     TO_CHAR(DATE_TRUNC('month', date), 'Month') AS month_name,
     ROUND(AVG(daily_annualized_yield_pct), 4) AS avg_yield_pct,
-    ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct), 4) AS median_yield_pct,
+    ROUND((PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY daily_annualized_yield_pct))::numeric, 4) AS median_yield_pct,
     COUNT(*) AS observation_count
 FROM marts.mart_daily_funding
 GROUP BY EXTRACT(MONTH FROM date), TO_CHAR(DATE_TRUNC('month', date), 'Month')
