+++ sql/equity-vs-crypto-perps/03_weekend_vs_weekday_crypto.sql
@@ -13,7 +13,7 @@ SELECT
     is_weekend,
     venue,
     ROUND(AVG(avg_rate_bps), 4) AS avg_rate_bps,
-    ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY avg_rate_bps), 4) AS median_rate_bps,
+    ROUND((PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY avg_rate_bps))::numeric, 4) AS median_rate_bps,
     ROUND(STDDEV(avg_rate_bps), 4) AS std_dev_bps,
     COUNT(*) AS observation_count
 FROM marts.mart_daily_funding
