+++ sql/equity-vs-crypto-perps/05_equity_venue_comparison.sql
@@ -11,7 +11,7 @@
 SELECT
     venue,
     ROUND(AVG(avg_rate_bps), 4) AS avg_rate_bps,
-    ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY avg_rate_bps), 4) AS median_rate_bps,
+    ROUND((PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY avg_rate_bps))::numeric, 4) AS median_rate_bps,
     ROUND(STDDEV(avg_rate_bps), 4) AS std_dev_bps,
     COUNT(DISTINCT symbol) AS total_symbols,
     COUNT(DISTINCT date) AS total_days
